+++ Classes/ormapper/TCategoryBase.m
@@ -299,7 +299,7 @@
  */
 - (void)getInsertSql:(NSMutableString *)s
 {
-    [s appendFormat:@"INSERT INTO Categories VALUES(%d", self.pid];
+    [s appendFormat:@"INSERT INTO Categories VALUES(%ld", (long)self.pid];
     [s appendString:@","];
     [s appendString:[self quoteSqlString:_name]];
     [s appendString:@","];
